requirements.txt
requests
-e git+https://github.com/LCBRU/lbrc_flask.git@main#egg=lbrc_flask
Flask
sqlalchemy-migrate
pymysql
